api/requirements.txt
fastapi
uvicorn
supabase
google-genai
httpx
apscheduler
python-dotenv